requirements.txt
pyrebase4
streamlit
streamlit_option_menu
firebase_admin
dash